fix reduce for array data in detector_based case
Was looking for a one-dim array of data in detector-based output.  But
since I fixed that (multiple freeways are now handled correctly) then I
have to change the reduce code here to properly expect a possible array
of arrays.

If not an array of arrays, make it so, then sum each entry.

test passes.